+++ src/main/java/woowacourse/auth/ui/AuthenticationInterceptor.java
@@ -2,18 +2,34 @@ package woowacourse.auth.ui;
 
 import org.springframework.web.servlet.HandlerInterceptor;
 import woowacourse.auth.support.AuthorizationExtractor;
+import woowacourse.auth.support.JwtTokenProvider;
+import woowacourse.shoppingcart.exception.TokenExpiredException;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
 public class AuthenticationInterceptor implements HandlerInterceptor {
 
-    private static final String ACCESS_TOKEN = "ACCESS_TOKEN";
+    private static final String TOKEN_PAYLOAD = "TOKEN_PAYLOAD";
+
+    private final JwtTokenProvider jwtTokenProvider;
+
+    public AuthenticationInterceptor(JwtTokenProvider jwtTokenProvider) {
+        this.jwtTokenProvider = jwtTokenProvider;
+    }
 
     @Override
     public boolean preHandle(final HttpServletRequest request, final HttpServletResponse response, Object handler) {
         final String accessToken = AuthorizationExtractor.extract(request);
-        request.setAttribute(ACCESS_TOKEN, accessToken);
+        validateTokenExpired(accessToken);
+
+        request.setAttribute(TOKEN_PAYLOAD, jwtTokenProvider.getPayload(accessToken));
         return true;
+    }
+
+    private void validateTokenExpired(final String token) {
+        if (!jwtTokenProvider.validateToken(token)) {
+            throw new TokenExpiredException();
+        }
     }
 }
